refactor: Moved buildedFormData to each type.file
* Added String+Constant.swift with String.newLine
