+++ OpenMarket/Constants/String+Constants.swift
@@ -1,0 +1,12 @@
+//
+//  File.swift
+//  OpenMarket
+//
+//  Created by kjs on 2021/08/24.
+//
+
+import Foundation
+
+extension String {
+    static let newLine = "\r\n"
+}
